backend/requirements.txt
django
django-ninja
psycopg[binary]
celery
redis
python-decouple
gunicorn
requests
django-cors-headers